+++ node/l1-bridge/src/event_queue.rs
@@ -28,11 +28,6 @@ impl EventQueue {
         self.queue.pop()
     }
 
-    /// Returns true if the queue is empty.
-    pub fn is_empty(&self) -> bool {
-        self.queue.is_empty()
-    }
-
     /// Waits until an event is available, then returns it.
     ///
     /// This is useful for consumers that want to block until events arrive.
@@ -45,13 +40,6 @@ impl EventQueue {
         }
     }
 
-    /// Waits for notification that events may be available.
-    ///
-    /// After this returns, call `pop()` to drain events.
-    pub async fn wait(&self) {
-        self.notify.notified().await;
-    }
-
     /// Drains all available events into a vector.
     pub fn drain(&self) -> Vec<L1Event> {
         let mut events = Vec::new();
@@ -59,28 +47,5 @@ impl EventQueue {
             events.push(event);
         }
         events
-    }
-
-    /// Waits until an event matching the predicate is found.
-    ///
-    /// Returns all collected events up to and including the matching event.
-    /// Use `tokio::time::timeout` to add a timeout if needed.
-    pub async fn wait_for<F>(&self, predicate: F) -> Vec<L1Event>
-    where
-        F: Fn(&L1Event) -> bool,
-    {
-        let mut collected = Vec::new();
-        loop {
-            // Drain any currently available events.
-            while let Some(event) = self.queue.pop() {
-                let matches = predicate(&event);
-                collected.push(event);
-                if matches {
-                    return collected;
-                }
-            }
-            // Wait for more events.
-            self.notify.notified().await;
-        }
     }
 }
