+++ node/test-suite/src/lib.rs
@@ -1,5 +1,7 @@
+mod event_queue_ext;
 mod l1_node;
 
+pub use event_queue_ext::EventQueueExt;
 pub use kaspa_consensus_core::{
     Hash,
     network::{NetworkId, NetworkType},
